Remove G* symbols and replace them with our own
This commit is a de-GLibification of Graphene's core API.

We don't use much of GLib any more: a couple of macros, mostly, for
validating input; plus, some convenience macros like MIN/MAX/CLAMP and
the likely/unlikely branch prediction hints.

We still need GLib for tests and benchmarks, and for the GObject types,
obviously, so this requirement won't go away unless you disable all of
those when configuring Graphene.